+++ datapkg/distribution/__init__.py
@@ -7,7 +7,6 @@ from datapkg import DatapkgException
 from datapkg.package import Package
 import datapkg.metadata as M
 from base import DistributionBase
-from simple import SimpleDistribution, IniBasedDistribution
 from python import PythonDistribution
 from jsondist import JsonDistribution